requirements/requirements.txt
cython
xtcocotools
wheel
matplotlib
opencv-python
sievedata
pathlib
hamer @ git+ssh://git@github.com/geopavlakos/hamer.git@main#egg=hamer
mmpose @ git+ssh://git@github.com/ViTAE-Transformer/ViTPose.git@main
hydra-core
omegaconf
webdataset